+++ src/pages/templates/EditPage.js
@@ -13,7 +13,7 @@ import { templateById } from 'src/selectors/templates';
 
 // Components
 import Form from './components/Form';
-import Editor from './components/Editor';
+import Editor from './components/AsyncEditor';
 import { DeleteModal } from 'src/components';
 import { Loading } from 'src/components';
 import { Page, Grid } from '@sparkpost/matchbox';
