+++ src/pages/templates/components/AsyncEditor.js
@@ -2,7 +2,7 @@ import loadable from 'react-loadable';
 import LoadableLoading from 'src/components/loading/LoadableLoading';
 
 export const LoadableEditor = loadable({
-  loader: () => import('./LineChart'),
+  loader: () => import('./Editor'),
   loading: LoadableLoading
 });
